+++ packages/titanium-chip/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.0.2](https://github.com/LeavittSoftware/titanium-elements/compare/@leavittsoftware/titanium-chip@4.0.1...@leavittsoftware/titanium-chip@4.0.2) (2022-04-29)
+
+**Note:** Version bump only for package @leavittsoftware/titanium-chip
+
+
+
+
+
 ## [4.0.1](https://github.com/LeavittSoftware/titanium-elements/compare/@leavittsoftware/titanium-chip@4.0.0...@leavittsoftware/titanium-chip@4.0.1) (2022-04-27)